After JoR meeting: remove FORMAT and IMPLICIT mentions; no Hollerith
JoR subgroup met today to discuss the updates to 25-114r0.

There was general agreement with the other changes already made.
There was agreement to remove the Fortran operators (separate PR).

This PR addresses the following decisions:
We should remove the references to FORMAT and IMPLICIT.
We should not insert anything about Hollerith.