Remove references to Fortran operators and constants
After discussion today, JoR subgroup agreed that we had no strong use
cases for adding Fortran relational and logical operators to the Fortran
preprocessor. Such operators do not appear with any frequency in our
corpus of example Fortran projects.

We agreed to limit the operators used to construct #if and #elif
expressions to those available in the C preprocessor for constructing
constant expressions.