Remove references to Fortran operators and constants (#39)
* Remove references to Fortran operators and constants

After discussion today, JoR subgroup agreed that we had no strong use
cases for adding Fortran relational and logical operators to the Fortran
preprocessor. Such operators do not appear with any frequency in our
corpus of example Fortran projects.

We agreed to limit the operators used to construct #if and #elif
expressions to those available in the C preprocessor for constructing
constant expressions.

Add a note that prior versions of requirements included Fortran operators
for use in `#if` and `#elif` expressions. In the absence of strong use cases,
these Fortran have been removed.

Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>

---------

Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>