Expand in only full-line comments; don't expand in alternative chars
We no longer try to discern "comment directives" for macro expansion.
A simpler proposal to expand in (full-line) comment lines is added.

A proposal to mention no token expansion in "alternative syntax for
character constants and edit descriptors" was added. (We're looking
at you, Hollerith.)

Add reference for the Fortran 2023 Interpretation document, 24-007,
as we use it to define "comment line" meaning a comment that contains
all the non-blank portion of a line.