Add initial (very rough) outline for specifications paper
Many subsections are not filled in yet.
At the end is a bunch of raw material that we
need to consider adding to the earlier sections.
This raw material comes from 25-114r2, Peter Klausler's
Preprocessing.md in Flang, and Berkely Lab's
assert library.There are probably other places
we need to pull from as well.