Resolve last comments from review: CHARACTER_STRING defn, #error exp
In the review comments and in the meeting today, we agreed
to limit the definition of CHARACTER_STRING to those enclosed
in double-quotes (as in C). This applies to the strings that
appear on #include and #line directives, and the values of
the predefined macros.

We also agreed that expansion of the token list on #warning
and #error directives shouldn't happen. Expansion there appears
to be counter to existing implementations.