4.1 Macro recognition and expansion: Add placeholder text
As discussed in our 2025-05-28 meeting, we want to have a placeholder in
section 4 as a backup plan in case schedule prevents us from finishing
a stand-alone section 4 in the next week.

This placeholder acknowledges the incomplete status, expresses our
technical intent and references the relevant sections of C 2023
(which are long and subtle, but I believe the technical content
really is almost exactly what we need to specify in final edits).

I've noted the few exceptions I'm aware of, where our previous
discussions have indicated at least a weak consensus for FPP
behavior to deliberately diverge from CPP.