ID: Fix Regex
The Regex for ID must allow a leading underscore, because we use
it in the grammar to stand for identifiers serving as either a macro
name or a FLM parameter name. Both allow a leading underscore, eg:

`#ifdef __LINE__`

`#define PLUS(_x, _y) (_x + _y)`

Both of these examples are fully valid and need to be permitted.